+++ src/colors.rs
@@ -1,0 +1,26 @@
+pub mod color {
+    pub type RGBColor = (u8, u8, u8);
+
+    pub const WHITE: RGBColor = (255, 255, 255);
+    pub const RED: RGBColor = (228, 3, 3);
+    pub const SINOPIA: RGBColor = (214, 41, 0);
+    pub const ORANGE: RGBColor = (255, 140, 0);
+    pub const ATOMIC_TANGERINE: RGBColor = (255, 155, 85);
+    pub const YELLOW: RGBColor = (255, 237, 0);
+    pub const GREEN: RGBColor = (0, 128, 38);
+    pub const MINT: RGBColor = (152, 232, 193);
+    pub const TURQUOISE: RGBColor = (38, 206, 170);
+    pub const TEAL: RGBColor = (7, 141, 112);
+    pub const MAYA_BLUE: RGBColor = (85, 205, 252);
+    pub const COBALT_BLUE: RGBColor = (123, 173, 226);
+    pub const MAJORELLE_BLUE: RGBColor = (80, 73, 204);
+    pub const BLUE: RGBColor = (0, 56, 168);
+    pub const INDIGO: RGBColor = (36, 64, 142);
+    pub const PURPLE: RGBColor = (61, 26, 120);
+    pub const MAUVE: RGBColor = (155, 79, 150);
+    pub const AMARANTH_PINK: RGBColor = (247, 168, 184);
+    pub const SUPER_PINK: RGBColor = (212, 97, 166);
+    pub const RUBY: RGBColor = (214, 2, 112);
+    pub const FLIRT: RGBColor = (165, 0, 98);
+    pub const VIOLET: RGBColor = (115, 41, 130);
+}
